refactor: Move pages to Auth and Main folders
Refactored the project structure to improve organization.  Moved authentication-related pages (`Login`, `Register`, `ForgotPassword`, `ResetPassword`, `VerifyEmail`) into an `Auth` folder and main application pages (`Home`, `Chat`, `Progress`, `Profile`, `NotFound`) into a `Main` folder.  No breaking changes; only file relocations.
