+++ frontend/src/pages/Auth/Register.tsx
